Remove changes from public API to preserve major version
