+++ src/Microsoft.ApplicationInsights/DataContracts/RequestTelemetry.cs
@@ -20,7 +20,7 @@
     /// method.
     /// <a href="https://go.microsoft.com/fwlink/?linkid=525722#trackrequest">Learn more</a>
     /// </remarks>
-    public sealed class RequestTelemetry : OperationTelemetry, ITelemetry, ISupportProperties, ISupportMetrics, ISupportSampling, IAiSerializableTelemetry
+    public sealed class RequestTelemetry : OperationTelemetry, ITelemetry, ISupportProperties, ISupportMetrics, ISupportSampling, ISupportAdvancedSampling, IAiSerializableTelemetry
     {
         internal new const string TelemetryName = "Request";
